Split the IP family stuff (eg IsIPv6String()) into its own file.
